Change the structure
