+++ app/Core/ReceiveQueue.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Comoyi\Hall\Objects;
+namespace Comoyi\Hall\Core;
 
 /**
  * 消息包接收队列
