+++ app/Handlers/Recv/GlobalMessageHandler.php
@@ -4,9 +4,9 @@ namespace Comoyi\Hall\Handlers\Recv;
 
 use Comoyi\Hall\Factory\SendMsgFactory;
 use Comoyi\Hall\Handlers\Recv\RecvHandler;
-use Comoyi\Hall\Models\CmdSend;
-use Comoyi\Hall\Models\CmdRecv;
-use Comoyi\Hall\Objects\Msg;
+use Comoyi\Hall\Cmd\CmdSend;
+use Comoyi\Hall\Cmd\CmdRecv;
+use Comoyi\Hall\Core\Msg;
 
 /**
  * 世界喇叭
